# optimize encoder