--- a/internal/priorityqueue/priorityqueue.go
+++ b/internal/priorityqueue/priorityqueue.go
@@ -17,10 +17,10 @@ type PriorityQueue[T any] struct {
 
 // New creates a new PriorityQueue, configured with a function that
 // compares the priorities of two items a and b; it should return a number > 0
-// if the priority of a is higher, 0 if the priorities are equal a number < 0
-// otherwise.
-func New[T any](cmp func(a, b T) int) *PriorityQueue[T] {
-	return &PriorityQueue[T]{cmp: cmp, items: make([]T, 1)}
+// if the priority of a is higher, 0 if the priorities are equal, and a number < 0 otherwise.
+// sizeHint sets the initial capacity of the queue; -1 means to use the default.
+func New[T any](sizeHint int, cmp func(a, b T) int) *PriorityQueue[T] {
+	return &PriorityQueue[T]{cmp: cmp, items: make([]T, 1, max(1, sizeHint+1))}
 }
 
 // Len returns the length (number of items) of the priority queue.