--- a/internal/priorityqueue/priorityqueue.go
+++ b/internal/priorityqueue/priorityqueue.go
@@ -1,5 +1,5 @@
-// Package priorityqueue provides a generic priority queue with Insert
-// and PopMax operations.
+// Package priorityqueue provides a generic priority queue with Insert,
+// PopMax, and RemoveFunc operations.
 package priorityqueue
 
 // PriorityQueue is a generic priority queue with a configurable comparison
@@ -44,8 +44,39 @@ func (pq *PriorityQueue[T]) PopMax() T {
 	maxItem := pq.items[1]
 	pq.items[1] = pq.items[len(pq.items)-1]
 	pq.items = pq.items[:len(pq.items)-1]
-	pq.siftdown()
+	pq.siftdown(1)
 	return maxItem
+}
+
+// RemoveFunc removes all elements for which rm returns true.
+func (pq *PriorityQueue[T]) RemoveFunc(rm func(T) bool) {
+	// This is effectively slices.DeleteFunc, but inlined because we start from index 1.
+	i := 1
+	for ; i < len(pq.items); i++ {
+		if rm(pq.items[i]) {
+			break
+		}
+	}
+	if i == len(pq.items) {
+		return // nothing to remove
+	}
+	for j := i + 1; j < len(pq.items); j++ {
+		if v := pq.items[j]; !rm(v) {
+			pq.items[i] = v
+			i++
+		}
+	}
+	// Clear the tail.
+	clear(pq.items[i:])
+	pq.items = pq.items[:i]
+	pq.rebuildHeap()
+}
+
+// rebuildHeap rebuilds the entire heap from scratch.
+func (pq *PriorityQueue[T]) rebuildHeap() {
+	for i := len(pq.items) / 2; i >= 1; i-- {
+		pq.siftdown(i)
+	}
 }
 
 func (pq *PriorityQueue[T]) siftup(n int) {
@@ -66,8 +97,7 @@ func (pq *PriorityQueue[T]) siftup(n int) {
 	}
 }
 
-func (pq *PriorityQueue[T]) siftdown() {
-	i := 1
+func (pq *PriorityQueue[T]) siftdown(i int) {
 	for {
 		c := 2 * i
 		if c >= len(pq.items) {