This is a grab-bag of optimizations.
I recommend reviewing commit-by-commit and rebasing instead of squashing.

Their cumulative effect, on my laptop, for an out-of-tree benchmark (sorry) is:

```
goos: darwin
goarch: arm64
pkg: bold.dev/tknz/gemma2b
cpu: Apple M3 Max
          │      a      │                  f                  │
          │   sec/op    │   sec/op     vs base                │
PureGo-16   3.494m ± 6%   2.642m ± 1%  -24.39% (p=0.000 n=15)

          │       a       │                  f                   │
          │     B/op      │     B/op      vs base                │
PureGo-16   1858.5Ki ± 0%   824.0Ki ± 0%  -55.66% (p=0.000 n=15)

          │       a       │                 f                  │
          │   allocs/op   │ allocs/op   vs base                │
PureGo-16   7574.000 ± 0%   4.000 ± 0%  -99.95% (p=0.000 n=15)
```

I also will understand if these are viewed as too intrusive/complicated for this codebase. :) I am happy to maintain a fork as needed.