@eliben 2024-08-27
Thanks, this is interesting. I'll find time to review these

I do wonder, however, about your use case. Can you elaborate where higher encoding performance is important to you, or is this just for fun? LLMs generally are pretty slow and in recent comparisons I've seen providers brag that their models reach something like 200 tokens/second (and that's probably on the beefiest HW). In comparison, unless I've messed up my benchmark, `go-sentencepiece` encodes at well over half a million tokens/sec, so it's hard for me to imagine a situation where this tokenization is a bottleneck.