@josharian 2024-08-27
> I'll find time to review these

Great, thanks. I made some decisions along the way with not 100% confidence. Pushback/questions are--always--welcome.

> Can you elaborate where higher encoding performance is important to you

As part of exploration and dataset preparation, I end up doing mass tokenization runs, so I really feel the performance--it directly impacts my iteration speed as I'm hacking around.

And I'm not working with frontier models (gemma, not gemini). And some of my work involves squeezing extra performance out, plus our use case is quite latency-sensitive.

I also generally have a habit of occasionally taking a day to crush any boxes I can in pprof. Cumulatively, that adds up.

Anyway, I put in this time to get to rough parity with the cgo implementation, at least for my use cases, so that I can switch to pure Go without guilt. Plus it was fun. :)