@josharian 2024-08-29
Will push up a revised copy once the questions above are answered. Sorry for the delay, been a busy couple of days.